+++ CHANGELOG.md
@@ -17,6 +17,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#1652])(https://github.com/open-telemetry/opentelemetry-python/pull/1652)
 - Rename Resource's `create_empty` to `get_empty`
   ([#1653])(https://github.com/open-telemetry/opentelemetry-python/pull/1653)
+- Renamed `BatchExportSpanProcessor` to `BatchSpanProcessor` and `SimpleExportSpanProcessor` to
+  `SimpleSpanProcessor`
+  ([#1656])(https://github.com/open-telemetry/opentelemetry-python/pull/1656)
 
 ## [0.18b0](https://github.com/open-telemetry/opentelemetry-python/releases/tag/v0.18b0) - 2021-02-16